Refactor: implementation of the ImpersonationService is extracted into an interface so it can be overriden when necessary
