+++ Source/Rhetos.WindowsAuthImpersonation/Abstractions/IImpersonationProvider.cs
@@ -1,0 +1,10 @@
+﻿namespace Rhetos.WindowsAuthImpersonation.Abstractions
+{
+    public interface IImpersonationProvider
+    {
+        string GetImpersonatedUserName();
+        string GetActualUserName();
+        void CheckImpersonatedUserPermissions(string impersonatedUser);
+        void SetImpersonatedUser(string impersonatedUser);
+    }
+}
